requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
groq==0.4.2
langchain-groq==0.0.6
sentence-transformers==2.5.1
faiss-cpu==1.7.4
pinecone-client==3.0.2
PyPDF2==3.0.1
numpy==1.24.3
requests==2.31.0
python-magic==0.4.27
typing-extensions==4.9.0
aiohttp==3.9.1
asyncio==3.4.3
python-multipart==0.0.6
pytest==7.4.3
black==23.11.0
pylint==3.0.2 